requirements.txt
Django==2.2.5
django-bootstrap-form==3.4
djangorestframework==3.10.3
html5lib==1.0.1
Pillow==6.2.0
pkg-resources==0.0.0
PyPDF2==1.26.0
pytz==2019.2
reportlab==3.5.28
six==1.12.0
sqlparse==0.3.0
webencodings==0.5.1
xhtml2pdf==0.2.3
xlwt==1.3.0
